Fronted/src/requirements.txt
fastapi
uvicorn
requests
googlesearch-python
youtube-search-python
pytube
beautifulsoup4
groq
